Uma/add bridge (#14)
* Ported bridge, tests pass

* Adding bridge stuff

* Initial moving around of .env and toml files

* Changed relayer to build with new config

* Changed operate to work with new config style

* Got cli to build

* Relaying finally works with new config system. Cleanup still neded

* Removed verify script

* Added bridge deploy and it works. Still need to add verification

* Added bridge deployment and it works

GitOrigin-RevId: 5906cb14e9e984dc6581e786febdc36f413bea9b

+++ test/amb/LightClientMock.sol
@@ -10,7 +10,7 @@ contract LightClientMock {
     mapping(uint256 => bytes32) public executionStateRoots;
     event HeadUpdate(uint256 indexed slot, bytes32 indexed root);
 
-    function setHeader(uint64 slot, bytes32 headerRoot) external {
+    function setHeader(uint256 slot, bytes32 headerRoot) external {
         headers[slot] = headerRoot;
         // NOTE that the stateRoot emitted here is not the same as the header root
         // in the real LightClient
